requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
Pillow==10.1.0
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
